requirements.txt
customtkinter>=5.2.0
google-generativeai>=0.8.0
yandex-ai-studio-sdk>=0.19.0
PyPDF2>=3.0.0
python-docx>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
